refactor(cli): extract workflow baseline handler
